requirements.txt
albucore==0.0.19
albumentations==1.4.20
annotated-types==0.7.0
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.0
cycler==0.12.1
efficientnet_pytorch==0.7.1
eval_type_backport==0.2.0
filelock==3.13.1
fonttools==4.54.1
fsspec==2024.2.0
huggingface-hub==0.26.1
idna==3.10
importlib_resources==6.4.5
Jinja2==3.1.3
joblib==1.4.2
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.2
mpmath==1.3.0
munch==4.0.0
networkx==3.2.1
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.3
pillow==10.2.0
pretrainedmodels==0.7.4
pydantic==2.9.2
pydantic_core==2.23.4
pydensecrf @ file:///home/conda/feedstock_root/build_artifacts/pydensecrf_1716289739392/work
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.13.1
segmentation-models-pytorch==0.3.4
six==1.16.0
stringzilla==3.10.7
sympy==1.12
threadpoolctl==3.5.0
timm==0.9.7
torch==2.4.1
torchaudio==2.4.1+cu124
torchsummary==1.5.1
torchvision==0.19.1+cu124
tqdm==4.66.5
triton==3.0.0
typing_extensions==4.9.0
tzdata==2024.2
urllib3==2.2.3
zipp==3.20.2
